# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CellDetailResponse": {
        "description": "Shared shape for `GET /spot` and `GET /cells/{id}`: today plus a\n7-day outlook, per species, with the full factor breakdown.",
        "properties": {
          "cell_id": {
            "title": "Cell Id",
            "type": "string"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          },
          "species": {
            "items": {
              "$ref": "#/components/schemas/SpeciesForecast"
            },
            "title": "Species",
            "type": "array"
          }
        },
        "required": [
          "cell_id",
          "lon",
          "lat",
          "place",
          "species"
        ],
        "title": "CellDetailResponse",
        "type": "object"
      },
...
      "ComuniResponse": {
        "properties": {
          "comuni": {
            "items": {
              "$ref": "#/components/schemas/Comune"
            },
            "title": "Comuni",
            "type": "array"
          }
        },
        "required": [
          "comuni"
        ],
        "title": "ComuniResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HotspotsResponse": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "hotspots": {
            "items": {
              "$ref": "#/components/schemas/Hotspot"
            },
            "title": "Hotspots",
            "type": "array"
          },
          "species": {
            "enum": [
              "porcini",
              "ovoli",
              "gallinacci",
              "combined"
            ],
            "title": "Species",
            "type": "string"
          }
        },
        "required": [
          "species",
          "date",
          "hotspots"
        ],
        "title": "HotspotsResponse",
        "type": "object"
      },
...
      "OutlookResponse": {
        "properties": {
          "area": {
            "$ref": "#/components/schemas/Area"
          },
          "baseline": {
            "$ref": "#/components/schemas/Baseline"
          },
          "good_share": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Good Share",
            "type": "number"
          },
          "issued": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "when the long-range forecast was fetched",
            "title": "Issued"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/OutlookPeriod"
            },
            "title": "Periods",
            "type": "array"
          },
          "rain_lead": {
            "$ref": "#/components/schemas/RainLead"
          },
          "rain_tilt": {
            "$ref": "#/components/schemas/RainTiltBands"
          },
          "season_to_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SeasonToDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "species": {
            "enum": [
              "porcini",
              "ovoli",
              "gallinacci"
            ],
            "title": "Species",
            "type": "string"
          },
          "window": {
            "$ref": "#/components/schemas/SeasonWindow"
          }
        },
        "required": [
          "species",
          "area",
          "issued",
          "good_share",
          "rain_lead",
          "rain_tilt",
          "baseline",
          "window",
          "season_to_date",
          "periods"
        ],
        "title": "OutlookResponse",
        "type": "object"
      },
      "Place": {
        "properties": {
          "comune": {
            "title": "Comune",
            "type": "string"
          },
          "nearest_place": {
            "title": "Nearest Place",
            "type": "string"
          }
        },
        "required": [
          "comune",
          "nearest_place"
        ],
        "title": "Place",
        "type": "object"
      },
...
      "ScoresResponse": {
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/GridCellScore"
            },
            "title": "Cells",
            "type": "array"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "species": {
            "enum": [
              "porcini",
              "ovoli",
              "gallinacci",
              "combined"
            ],
            "title": "Species",
            "type": "string"
          }
        },
        "required": [
          "species",
          "date",
          "cells"
        ],
        "title": "ScoresResponse",
        "type": "object"
      },
      "SeasonMapResponse": {
        "description": "One season on the map: good days per woodland cell, and the comuni ranked by them.",
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CellSeason"
            },
            "title": "Cells",
            "type": "array"
          },
          "complete": {
            "title": "Complete",
            "type": "boolean"
          },
          "comuni": {
            "items": {
              "$ref": "#/components/schemas/ComuneSeason"
            },
            "title": "Comuni",
            "type": "array"
          },
          "good_score": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Good Score",
            "type": "number"
          },
          "species": {
            "enum": [
              "porcini",
              "ovoli",
              "gallinacci",
              "combined"
            ],
            "title": "Species",
            "type": "string"
          },
          "through": {
            "format": "date",
            "title": "Through",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "species",
          "complete",
          "through",
          "good_score",
          "cells",
          "comuni"
        ],
        "title": "SeasonMapResponse",
        "type": "object"
      },
...
      "SeasonsResponse": {
        "properties": {
          "area": {
            "$ref": "#/components/schemas/Area"
          },
          "baseline": {
            "$ref": "#/components/schemas/Baseline"
          },
          "good_score": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "Good Score",
            "type": "number"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/SeasonSummary"
            },
            "title": "Seasons",
            "type": "array"
          },
          "species": {
            "enum": [
              "porcini",
              "ovoli",
              "gallinacci",
              "combined"
            ],
            "title": "Species",
            "type": "string"
          }
        },
        "required": [
          "species",
          "area",
          "good_score",
          "baseline",
          "seasons"
        ],
        "title": "SeasonsResponse",
        "type": "object"
      },
...
      "SightingsResponse": {
        "properties": {
          "counts": {
            "items": {
              "$ref": "#/components/schemas/SightingCount"
            },
            "title": "Counts",
            "type": "array"
          },
          "since": {
            "format": "date",
            "title": "Since",
            "type": "string"
          },
          "species": {
            "enum": [
              "porcini",
              "ovoli",
              "gallinacci"
            ],
            "title": "Species",
            "type": "string"
          }
        },
        "required": [
          "species",
          "since",
          "counts"
        ],
        "title": "SightingsResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Fruiting-conditions scores, cell detail, hotspots and sightings for wild edible mushrooms in Tuscany. Scores are a 0-1 conditions index, never a probability or an edibility claim.",
    "title": "mushma API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cells/{cell_id}": {
      "get": {
        "operationId": "get_cell_cells__cell_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "cell_id",
            "required": true,
            "schema": {
              "title": "Cell Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown cell id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score, 7-day outlook and factor breakdown for every species in one cell"
      }
    },
    "/comuni": {
      "get": {
        "operationId": "get_comuni_comuni_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComuniResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "history not built yet"
          }
        },
        "summary": "Comuni with woodland, for the season and outlook views"
      }
    },
...
    "/history/season/{year}": {
      "get": {
        "operationId": "get_season_map_history_season__year__get",
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "title": "Year",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "species",
            "required": true,
            "schema": {
              "enum": [
                "porcini",
                "ovoli",
                "gallinacci",
                "combined"
              ],
              "title": "Species",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonMapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "season not stored"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "history not built yet"
          }
        },
        "summary": "One season on the map: good days per woodland cell, comuni ranked by them"
      }
    },
    "/history/seasons": {
      "get": {
        "operationId": "get_seasons_history_seasons_get",
        "parameters": [
          {
            "description": "ISTAT comune code (see /comuni); omit for all of Tuscany",
            "in": "query",
            "name": "comune",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISTAT comune code (see /comuni); omit for all of Tuscany",
              "title": "Comune"
            }
          },
          {
            "in": "query",
            "name": "species",
            "required": false,
            "schema": {
              "default": "combined",
              "enum": [
                "porcini",
                "ovoli",
                "gallinacci",
                "combined"
              ],
              "title": "Species",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeasonsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown comune"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "history not built yet"
          }
        },
        "summary": "Every stored season for Tuscany or a comune: good days, weather vs normal, sightings"
      }
    },
    "/hotspots": {
      "get": {
        "operationId": "get_hotspots_hotspots_get",
        "parameters": [
          {
            "in": "query",
            "name": "species",
            "required": true,
            "schema": {
              "enum": [
                "porcini",
                "ovoli",
                "gallinacci",
                "combined"
              ],
              "title": "Species",
              "type": "string"
            }
          },
          {
            "description": "defaults to today, Europe/Rome",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "defaults to today, Europe/Rome",
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "date outside the served window"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ranked clusters of high-scoring cells, with nearby recent sightings"
      }
    },
    "/outlook": {
      "get": {
        "operationId": "get_outlook_outlook_get",
        "parameters": [
          {
            "in": "query",
            "name": "species",
            "required": true,
            "schema": {
              "enum": [
                "porcini",
                "ovoli",
                "gallinacci"
              ],
              "title": "Species",
              "type": "string"
            }
          },
          {
            "description": "ISTAT comune code (see /comuni); omit for all of Tuscany",
            "in": "query",
            "name": "comune",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISTAT comune code (see /comuni); omit for all of Tuscany",
              "title": "Comune"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown comune"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "history not built yet"
          }
        },
        "summary": "The season so far and an outlook (not a forecast) for the weeks and months ahead"
      }
    },
    "/scores": {
      "get": {
        "operationId": "get_scores_scores_get",
        "parameters": [
          {
            "in": "query",
            "name": "species",
            "required": true,
            "schema": {
              "enum": [
                "porcini",
                "ovoli",
                "gallinacci",
                "combined"
              ],
              "title": "Species",
              "type": "string"
            }
          },
          {
            "description": "defaults to today, Europe/Rome",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "defaults to today, Europe/Rome",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoresResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "date outside the served window"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Whole-region grid of conditions scores for one species (or combined)"
      }
    },
    "/sightings": {
      "get": {
        "operationId": "get_sightings_sightings_get",
        "parameters": [
          {
            "in": "query",
            "name": "species",
            "required": true,
            "schema": {
              "enum": [
                "porcini",
                "ovoli",
                "gallinacci"
              ],
              "title": "Species",
              "type": "string"
            }
          },
          {
            "description": "defaults to one year ago",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "defaults to one year ago",
              "title": "Since"
            }
          },
          {
            "description": "last day included; defaults to no limit",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "last day included; defaults to no limit",
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SightingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public sighting counts per cell (never coordinates)"
      }
    },
    "/species": {
      "get": {
        "operationId": "get_species_species_get",
        "parameters": [
          {
            "description": "ISTAT comune code (see /comuni); omit for all of Tuscany",
            "in": "query",
            "name": "comune",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISTAT comune code (see /comuni); omit for all of Tuscany",
              "title": "Comune"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlausibleSpeciesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "unknown comune"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "history not built yet"
          }
        },
        "summary": "Plausible species for Tuscany or a comune: habitat fit and good days per season, per species and taxon"
      }
    },
    "/spot": {
      "get": {
        "operationId": "get_spot_spot_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "maximum": 90,
              "minimum": -90,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "maximum": 180,
              "minimum": -180,
              "title": "Lon",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score, 7-day outlook and factor breakdown for every species at a point"
      }
    },
    "/status": {
      "get": {
        "operationId": "get_status_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "description": "the pipeline has never scored anything yet"
          }
        },
        "summary": "Data freshness: the latest scored day, when it was generated, and the rules version"
      }
    }
  }
}